* Add makefile and changes necessary for building on linux. * Split source into headers and separate sources.
